third_party/blink/web_tests/platform/generic/touchadjustment/iframe-expected.txt

Test touch-adjustment on links in an iframe. Making sure we iframe position is correctly adjusted for.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test fat direct touches.
PASS adjustedNode.id is "a1"
PASS adjustedNode.id is "a2"
PASS successfullyParsed is true

TEST COMPLETE
